feat: add session API/UI modules and tests
