+++ public/js/__tests__/sessionApi.test.js
@@ -1,0 +1,54 @@
+jest.mock('../sessionManager.js', () => ({
+  getAuthToken: () => 'token'
+}));
+
+const realFetch = global.fetch;
+
+afterEach(() => {
+  jest.resetModules();
+  localStorage.clear();
+  if(realFetch){
+    global.fetch = realFetch;
+  } else {
+    delete global.fetch;
+  }
+  delete global.io;
+});
+
+describe('sessionApi', () => {
+  test('getSessions fetches and stores sessions', async () => {
+    const mockSessions = [
+      { id: '1', name: 'A' },
+      { id: '2', name: 'B', archived: 1 }
+    ];
+    global.fetch = jest.fn(() => Promise.resolve({ ok: true, json: () => Promise.resolve(mockSessions) }));
+    const { getSessions } = require('../sessionApi.js');
+    const list = await getSessions();
+    expect(list).toEqual([
+      { id: '1', name: 'A', archived: false },
+      { id: '2', name: 'B', archived: true }
+    ]);
+    expect(JSON.parse(localStorage.getItem('trauma_sessions'))[1].archived).toBe(true);
+  });
+
+  test('saveSessions posts list', async () => {
+    global.fetch = jest.fn(() => Promise.resolve({ ok: true }));
+    const { saveSessions } = require('../sessionApi.js');
+    const list = [{ id: '1', name: 'A', archived: false }];
+    await saveSessions(list);
+    expect(global.fetch).toHaveBeenCalledWith('/api/sessions', expect.objectContaining({ method: 'PUT' }));
+  });
+
+  test('connectSocket registers callbacks', () => {
+    const handlers = {};
+    global.io = jest.fn(() => ({
+      on: (evt, cb) => { handlers[evt] = cb; },
+      disconnect: jest.fn()
+    }));
+    const { connectSocket } = require('../sessionApi.js');
+    const onSessions = jest.fn();
+    connectSocket({ onSessions });
+    handlers.sessions([1,2]);
+    expect(onSessions).toHaveBeenCalledWith([1,2]);
+  });
+});
